+++ lib/liquidbook.rb
@@ -5,10 +5,13 @@ require_relative "liquidbook/version"
 require_relative "liquidbook/config"
 require_relative "liquidbook/schema_parser"
 require_relative "liquidbook/param_parser"
+require_relative "liquidbook/template_analyzer"
+require_relative "liquidbook/filter_type_map"
+require_relative "liquidbook/parameter_merger"
 require_relative "liquidbook/mock_data"
 require_relative "liquidbook/filters/shopify_filters"
 require_relative "liquidbook/tags/section_tag"
-require_relative "liquidbook/tags/render_tag"
+require_relative "liquidbook/theme_file_system"
 require_relative "liquidbook/pid_manager"
 require_relative "liquidbook/theme_renderer"
 require_relative "liquidbook/server/app"
@@ -43,7 +46,6 @@ module Liquidbook
       Liquid::Environment.build do |e|
         e.register_filter(Filters::ShopifyFilters)
         e.register_tag("section", Tags::SectionTag)
-        e.register_tag("render", Tags::RenderTag)
       end
     end
   end